services/api/pyproject.toml
[project]
name = "kops-api"
version = "0.1.0"
description = "KnowledgeOps AI — FastAPI backend (api + shared modules used by worker)."
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "redis==5.0.8",
    "argon2-cffi==23.1.0",
    "pyjwt==2.9.0",
    "python-multipart==0.0.12",
    "httpx==0.27.2",
    "chromadb-client==0.5.23",
    "google-generativeai==0.8.3",
    "celery==5.4.0",
    "structlog==24.4.0",
    "python-json-logger==2.0.7",
    "email-validator==2.2.0",
    "pypdf==5.1.0",
    "pdfminer.six==20240706",
    "filetype==1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "ruff==0.6.9",
    "mypy==1.11.2",
    "aiosqlite==0.20.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "ASYNC", "S", "RUF"]
ignore = ["S101", "B008"]

[tool.ruff.lint.per-file-ignores]
"app/tests/**" = ["S105", "S106"]
"migrations/**" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["app/tests"]
addopts = "-q --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
